Restructure integration test folder, split utilities from tests
