+++ integration-tests/auth.ts
@@ -1,7 +1,7 @@
 import { adminClient, createUserClient, defaultClient, test } from './base';
 import { pupilOne } from './user';
 import assert from 'assert';
-import { assertUserReceivedNotification, createMockNotification } from './notifications';
+import { assertUserReceivedNotification, createMockNotification } from './base/notifications';
 import { randomBytes } from 'crypto';
 
 void test('Token Login', async () => {
